femu: fix the AER count, Zone Append limit, KV MDTS and vector masking
AERL 255 allows 256 outstanding Async Event Requests, but the count of
held requests was 8 bits wide: it wrapped at the 256th, and the next
request was taken too and overwrote the first.

A ZASL of 0 means the Zone Append limit is MDTS, and an MDTS of 0 means
no limit, yet appends were capped at one page. Key value Store and
Retrieve never checked MDTS; they now bound the value they move.

Masking an MSI-X vector detached the irqfd of only the first completion
queue on it, while unmasking attaches every one, so the others kept
raising interrupts through a masked vector. That path needs KVM, so no
qtest covers it.

+++ hw/femu/nvme.h
@@ -1947,7 +1947,7 @@ typedef struct FemuCtrl {
     uint8_t     intc;
     uint8_t     intc_thresh;
     uint8_t     intc_time;
-    uint8_t     outstanding_aers;
+    uint16_t    outstanding_aers;   /* up to aerl + 1, which can be 256 */
     uint8_t     temp_warn_issued;
     uint64_t    num_errors;
     uint8_t     cqes_pending;
